+++ examples/fallbacks-1.ts
@@ -1,4 +1,5 @@
-import * as flow from 'dropflow/with-parse.js';
+import * as flow from 'dropflow';
+import parse from 'dropflow/parse.js';
 import fs from 'fs';
 import {createCanvas} from 'canvas';
 import {registerFontAsset} from '../assets/register.js';
@@ -8,7 +9,7 @@ registerFontAsset('Ramabhadra/Ramabhadra-Regular.ttf');
 registerFontAsset('Arimo/Arimo-Regular.ttf');
 registerFontAsset('Arimo/Arimo-Italic.ttf');
 
-const rootElement = flow.parse(`
+const rootElement = parse(`
   <div x-dropflow-log style="zoom: 2; font-family: Arimo; background-color: white; text-align: center;">
     <span style="font-style: italic; font-size: 0.5em;">English:</span> Welcome<br>
     <span style="font-style: italic; font-size: 0.5em;">Arabic:</span> أهلاً و سهلاً<br>