+++ test/font.spec.js
@@ -1,7 +1,8 @@
 //@ts-check
 import {expect} from 'chai';
 import fs from 'node:fs';
-import * as flow from 'dropflow/with-parse.js';
+import * as flow from 'dropflow';
+import parse from 'dropflow/parse.js';
 import {registerFontAsset, unregisterFontAsset} from '../assets/register.js';
 import {getCascade, fonts, FontFace, createFaceFromTables} from '../src/text-font.js';
 import {getOriginStyle, createStyle, createDeclaredStyle} from '../src/style.js';
@@ -89,10 +90,10 @@ describe('Fonts', function () {
     });
 
     it('looks up based on script when nothing is specified', function () {
-      const bc1 = flow.generate(flow.parse('هل تتحدث لغة أخرى بجانب العربية؟'));
+      const bc1 = flow.generate(parse('هل تتحدث لغة أخرى بجانب العربية؟'));
       flow.layout(bc1);
       expect(bc1.children[0].paragraph.wholeItems[0].face.family).to.equal('Cairo');
-      const bc2 = flow.generate(flow.parse('Do you speak another language besides Arabic?'));
+      const bc2 = flow.generate(parse('Do you speak another language besides Arabic?'));
       flow.layout(bc2);
       expect(bc2.children[0].paragraph.wholeItems[0].face.family).to.equal('Arimo');
     });