+++ test/text.spec.js
@@ -1,6 +1,7 @@
 //@ts-check
 import {expect} from 'chai';
-import * as flow from 'dropflow/with-parse.js';
+import * as flow from 'dropflow';
+import parse from 'dropflow/parse.js';
 import {registerFontAsset, unregisterFontAsset} from '../assets/register.js';
 import {G_ID, G_AX, G_SZ} from '../src/layout-text.js';
 import paintBlockContainer from '../src/paint.js';
@@ -11,7 +12,7 @@ const log = new Logger();
 
 function setupLayoutTests() {
   this.layout = function (html) {
-    this.rootElement = flow.parse(html);
+    this.rootElement = parse(html);
     this.blockContainer = flow.generate(this.rootElement);
     flow.layout(this.blockContainer);
     this.get = function (...args) {