+++ CHANGELOG.md
@@ -13,6 +13,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 * Removed `getRootStyle`
 * paintToCanvas no longer has a density argument. Use the zoom CSS property instead.
 * Changed the font registration API to match `document.fonts` in web browsers. Instead of `registerFont`, import `fonts` and `FontFace`. See the README for more details.
+* `parse` is now an individual file without the rest of the API. Change `import * as flow from 'dropflow/with-parse.js'` to `import * as flow from 'dropflow'` and `import parse from 'dropflow/parse.js'`
 
 ### Added
 * Added support for the `zoom` property